requirements.txt
flask
flask-sock
sounddevice
numpy
websocket-client
